Added a note to disable the console appearing.
Since the project cannot be built without tools enabled, we cannot create a truly "release" build of the game. At best, we can only make a "release_debug" build. This causes problems on Windows where the console only disappears in "release" builds.

This note will be in full detail later, I'm just putting it here to remind myself how to do it.