Added the ability to import glTF scenes.
Also realised during this that "assets" can have more than one file.
So, the code to import has been split into two: one function to copy
a file over (which also checks the .md5 file to see if it has changed),
and another function to create the .stex and .import files.