Fixed ErrorReporter not preserving UTF-8 encoded strings.
So, as per usual, it wasn't the engine's fault, it was mine. Again.

In attempting to fix drwhut/tabletop-club#291, I had thought there
was a flaw in how the engine handled errors and warnings, since the
error message was converted from a Godot string to a C string, I
thought that any UTF-8 encoding was lost in that conversion. I was
wrong - the data was still there, but when it reached the
ErrorReporter I had used the default String() constructor, not the
static String::utf8() function, which handles UTF-8 strings.

AAAAAAAAAAAAAAAAA